+++ test/test_functional2.cpp
@@ -28,8 +28,7 @@
 #include <jau/test/catch2_ext.hpp>
 
 #include "functional2.hpp"
-static std::string impl_name = "test/functional2.hpp";
 
 #define FUNCTIONAL_PROVIDED 1
-#define FUNCTIONAL_BROKEN_COPY_WITH_MUTABLE_LAMBDA 1
+#define FUNCTIONAL_IMPL 2
 #include "test_functional.hpp"